+++ Demo/Application/Features/PayPalWebCheckoutViewController.swift
@@ -250,14 +250,8 @@ class PayPalWebCheckoutViewController: PaymentButtonBaseViewController {
         sender.setTitle("Processing...", for: .disabled)
         sender.isEnabled = false
         
-        guard let userEmail = emailTextField.text, !userEmail.isEmpty else {
-            self.progressBlock("Email cannot be nil for App Switch flow")
-            sender.isEnabled = true
-            return
-        }
-
         let request = BTPayPalCheckoutRequest(
-            userAuthenticationEmail: userEmail,
+            userAuthenticationEmail: emailTextField.text,
             enablePayPalAppSwitch: true,
             amount: "10.00"
         )
@@ -277,15 +271,9 @@ class PayPalWebCheckoutViewController: PaymentButtonBaseViewController {
     @objc func tappedPayPalAppSwitchForVault(_ sender: UIButton) {
         sender.setTitle("Processing...", for: .disabled)
         sender.isEnabled = false
-        
-        guard let userEmail = emailTextField.text, !userEmail.isEmpty else {
-            self.progressBlock("Email cannot be nil for App Switch flow")
-            sender.isEnabled = true
-            return
-        }
 
         let request = BTPayPalVaultRequest(
-            userAuthenticationEmail: userEmail,
+            userAuthenticationEmail: emailTextField.text,
             enablePayPalAppSwitch: true
         )