+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 * BraintreePayPal
   * Bug fix: remove internal code signing from framework
+  * Make `userAuthenticationEmail` property optional for all App Switch flows
 * BraintreeThreeDSecure
   * Bug fix: remove internal code signing from framework